LoLockRC added
The results are not good as expected because in the integration of Lorenz's equations one need the 2 others components that are not locked.
Moreover, after stop the locking, the "Cuomo's synchronisation" occurs and the Lorenz locked turn back on the dynamic of the not locked system.

We need 2 others trained reservoir to lock the 3 components at the same time.

In the RCLockLo file the problem don't exist because we integrate the full system and locked the trained reservoir on the x component.